// src/types/partials/QuestionnaireItem.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TUri } from '../simpleTypes/Uri';
import { TCoding } from '../partials/Coding';
import { TQuestionnaireEnableWhen } from '../partials/QuestionnaireEnableWhen';
import { TInt } from '../simpleTypes/Int';
import { TCanonical } from '../simpleTypes/Canonical';
import { TQuestionnaireAnswerOption } from '../partials/QuestionnaireAnswerOption';
import { TQuestionnaireInitial } from '../partials/QuestionnaireInitial';

export type TQuestionnaireItem = {
  id?: string;
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  linkId: string;
  definition?: TUri;
  code?: TCoding[];
  prefix?: string;
  text?: string;
  type: string;
  enableWhen?: TQuestionnaireEnableWhen[];
  enableBehavior?: string;
  required?: boolean;
  repeats?: boolean;
  readOnly?: boolean;
  maxLength?: TInt;
  answerValueSet?: TCanonical;
  answerOption?: TQuestionnaireAnswerOption[];
  initial?: TQuestionnaireInitial[];
  item?: TQuestionnaireItem[];
};